+++ lib/services/flutter_service.dart
@@ -8,15 +8,12 @@ class FlutterService {
   /// Default constructor for [FlutterService].
   FlutterService();
 
-  /// Adds core Firebase and FlutterFire dependencies to the project.
+  /// Adds core Firebase dependencies to the project.
   Future<void> addDeps() async {
     await run('flutter', [
       'pub',
       'add',
       'firebase_core',
-      'firebase_auth',
-      'cloud_firestore',
-      'firebase_messaging'
     ]);
   }